add iris example
